+++ src/Domain/Search/ContentQueryBuilderInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+
+namespace Proximum\Vimeet\Domain\Search;
+
+interface ContentQueryBuilderInterface
+{
+    public function getQuery(string $search, string $locale): array;
+}