+++ src/Domain/Template/TemplateObject/Tag.php
@@ -15,10 +15,12 @@ use Proximum\Vimeet\Domain\Template\TemplateObject;
 class Tag extends TemplateObject implements SearchableObjectInterface
 {
     /**
+     * @param string|null $locale
+     *
      * @return array|string
      */
-    public function getSearchableContent()
+    public function getSearchableContent(?string $locale = null)
     {
-        return $this->getTaggedDataContent();
+        return $this->getTaggedDataContent($locale);
     }
 }