+++ src/Infrastructure/Adapter/SheetSearchQueryBuilder.php
@@ -14,7 +14,6 @@ use Elastica\Filter\Exists;
 use Elastica\Query\BoolQuery;
 use Elastica\Query\Filtered;
 use Elastica\Query\Match;
-use Elastica\Query\MultiMatch;
 use Elastica\Query\Nested;
 use Elastica\Query\Range;
 use Elastica\Query\Term;
@@ -37,18 +36,11 @@ use Proximum\Vimeet\Domain\Template\TemplateObject\Nomenclature;
 use Proximum\Vimeet\Domain\Type\TypeInterface;
 use Proximum\Vimeet\Domain\View\Catalog\CategoryView;
 use Proximum\Vimeet\Domain\View\Catalog\OrganizationCategoryView;
-use Proximum\Vimeet\Infrastructure\Elastica\AvailableLocales;
+use Proximum\Vimeet\Infrastructure\Elastica\QueryBuilder\ContentQueryBuilder;
 use Proximum\Vimeet\Infrastructure\Elastica\QueryBuilder\NomenclatureQueryBuilder;
 
 class SheetSearchQueryBuilder
 {
-    const BOOSTER_SHEET_NAME      = 5;
-    const BOOSTER_LOCALE_CONTENT  = 3;
-    const BOOSTER_DEFAULT_CONTENT = 2;
-
-    // Percentage content minimum should match
-    const CONTENT_MINIMUM_SHOULD_MATCH = 90;
-
     /**
      * @var BoolQuery
      */
@@ -230,26 +222,8 @@ class SheetSearchQueryBuilder
 
         $search = str_replace('|', ' ', $filters[SearchFields::FILTER_CONTENT]);
 
-        // Boost sheetname and content
-        $fields = [
-            sprintf('sheetName^%s', $this->initialBooster * self::BOOSTER_SHEET_NAME),
-            sprintf('content^%s', $this->initialBooster * self::BOOSTER_DEFAULT_CONTENT),
-        ];
-
-        if (in_array($this->locale, AvailableLocales::getAvailableLocalesForContent())) {
-            // If locale field is available
-            $fields[] = sprintf('content_%s^%s', $this->locale, $this->initialBooster * self::BOOSTER_LOCALE_CONTENT);
-        }
-
-        $multiMatch = new MultiMatch();
-        $multiMatch
-            ->setMinimumShouldMatch(self::CONTENT_MINIMUM_SHOULD_MATCH . '%')
-            ->setFields($fields)
-            ->setType(MultiMatch::TYPE_CROSS_FIELDS)
-            ->setQuery($search)
-        ;
-
-        $this->query->addMust($multiMatch);
+        $builder = new ContentQueryBuilder();
+        $this->query->addMust($builder->getElasticaQuery($search, $this->locale));
     }
 
     /**